Turn store::complete and store::enqueue into module dirs.
Co-locates the batchers and helpers with the modules that ultimately own
them. I don't love how much code now sits in mod.rs but we'll clean that
up later.
