+++ src/store/mod.rs
@@ -2,11 +2,9 @@
 // Licensed under the Business Source License 1.1. See LICENSE file for details.
 
 mod complete;
-mod complete_batcher;
 mod cron;
 mod delete;
 mod enqueue;
-mod enqueue_batcher;
 mod fail;
 mod group_committer;
 mod options;
